requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
opencv-python>=4.7.0
tqdm>=4.65.0
wandb>=0.15.0
python-box>=7.0.1
pyyaml>=6.0.0
scikit-learn>=1.2.0